Implement support for struct plugin as datastore
References database supports the data plugin database as storage. This
commit implements a similar functionality using the string plugin. If
a page within the 'reference-db-namespace' name space contains a page
with the struct-table block, we load the data from the specified schema
and construct references using the data contained in that database.
Users can edit the schema and its data as normal. The code has only been
tested with page data and not global or serial data. The commit creates
a dependency on struct - the code does not verify if the plugin is
installed or not.

Resolves: